+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Upgrading to Jitsi version 2.0.5390-3
 * Moving to common Makefile
 * Upgrading VPC common library (changes parameter names)
+* Disabling recording and live stream buttons
+* Disable default apache site
 
 1.0.4
 -----